modules/locationgrpc/requirements.txt
grpcio-tools==1.41.0
protobuf==3.18.1
six==1.16.0